+++ apps/web/src/services/auth/components/SignInRightContainer.vue
@@ -83,7 +83,7 @@ import { store } from '@/store';
 
 import config from '@/lib/config';
 
-import { AUTH_ROUTE } from '@/services/auth/routes/route-config';
+import { AUTH_ROUTE } from '@/services/auth/routes/route-constant';
 
 export default {
     name: 'SignInRightContainer',
