+++ apps/web/src/services/auth/components/ValidationEmailCollapsibleContents.vue
@@ -47,7 +47,7 @@ import { store } from '@/store';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
-import { AUTH_ROUTE } from '@/services/auth/routes/route-config';
+import { AUTH_ROUTE } from '@/services/auth/routes/route-constant';
 
 
 const { query } = SpaceRouter.router.currentRoute;
