+++ apps/web/src/services/auth/pages/PasswordPage.vue
@@ -114,7 +114,7 @@ import CenteredPageLayout from '@/common/modules/page-layouts/CenteredPageLayout
 
 import PasswordForm from '@/services/auth/components/PasswordForm.vue';
 import { PASSWORD_STATUS } from '@/services/auth/constants/password-constant';
-import { AUTH_ROUTE } from '@/services/auth/routes/route-config';
+import { AUTH_ROUTE } from '@/services/auth/routes/route-constant';
 import type { PasswordFormExpose } from '@/services/auth/types/password-type';
 
 interface Props {
